requirements.txt
requests[socks]
rich
python-dotenv
stem
dnspython
